Moving PkgResolution model to model.dart
