+++ test/maintenance_test.dart
@@ -6,7 +6,6 @@ import 'dart:convert';
 
 import 'package:pana/src/maintenance.dart';
 import 'package:pana/src/model.dart';
-import 'package:pana/src/pkg_resolution.dart';
 import 'package:pana/src/pubspec.dart';
 import 'package:test/test.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;
